reset the whole esy thing, try to clean the build stuff

+++ ml-clean.sh
@@ -1,9 +1,14 @@
 
 
-for f in $(find . -iname '*.ml'); do
-  [ -f "$f" ] || continue
+for d in native lib ; do
+  [ -d "$d" ] || continue
 
-  echo delete $f
-  rm -f $f
+  for f in $(find $d -iname '*.ml'); do
+    [ -f "$f" ] || continue
+
+    echo delete $f
+    rm -f $f
+  done
+
 done
